pyproject.toml
[project]
name = "supervised-graphs"
version = "0.1.0"
description = ""
authors = [
    { name = "c72bcbf4", email = "c72bcbf4@gmail.com" }
]
requires-python = ">=3.12"

dependencies = [
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "torch-geometric>=2.6.1",
    "torchmetrics>=1.7.1",
    "torchrl>=0.8.0",
    "pytorch-lightning>=2.5.1.post0",
    "line-profiler>=4.2.0",
    "networkx>=3.4.2",
    "tqdm>=4.67.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "scipy>=1.15.2",
    "pandas>=2.2.3",
    "imageio[ffmpeg]>=2.37.0",
    "line-profiler-pycharm>=1.2.0",
    "black>=25.1.0",
    "hydra-core>=1.3.2",
    "submitit>=1.5.2",
    "loguru>=0.7.3",
    "pynvml>=12.0.0",
    "rdkit>=2024.9.6"
]
